+++ tgcalls/src/NativeInstance.h
@@ -68,7 +68,7 @@ public:
     void setAudioOutputDevice(std::string id) const;
     void setAudioInputDevice(std::string id) const;
 
-    void setVideoCapture(std::function<std::string()>, int, int, int);
+    void setVideoCapture(std::function<std::string()>, float, int, int);
 
     void receiveSignalingData(std::vector<uint8_t> &data) const;
     void setJoinResponsePayload(std::string const &) const;
